+++ src/ZeroDowntimeMigrationTimeoutException.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Tpetry\PostgresqlEnhanced;
+
+use Exception;
+
+class ZeroDowntimeMigrationTimeoutException extends Exception
+{
+}